refactor: move common to a module
