+++ tensilelite/Tensile/Common/Common.py
@@ -22,28 +22,23 @@
 #
 ################################################################################
 
-from . import __version__
-from . import Parallel
-from .TensileInstructions import getGfxName, TensileInstructions
+from enum import Enum
 from collections import OrderedDict
 from copy import deepcopy
-from typing import Tuple
 
 import math
 import os.path
 import subprocess
 import sys
+import functools
 import time
 import re
 
-
-IsaVersion = Tuple[int, int, int]
-
+from Tensile.TensileInstructions import getGfxName, TensileInstructions
+from Tensile import __version__
 
 startTime = time.time()
 
-ParallelMap = Parallel.ParallelMap
-ParallelMap2 = Parallel.ParallelMap2
 # print level
 # 0 - user wants no printing
 # 1 - user wants limited prints
@@ -235,9 +230,6 @@ globalParameters["CurrentISA"] = (0,0,0)
 globalParameters["AMDGPUArchPath"] = None      # /opt/rocm/llvm/bin/amdgpu-arch
 globalParameters["ROCmAgentEnumeratorPath"] = None      # /opt/rocm/bin/rocm_agent_enumerator
 globalParameters["ROCmSMIPath"] = None                  # /opt/rocm/bin/rocm-smi
-globalParameters["IndexChars"] =  "IJKLMNOPQRSTUVWXYZ"  # which characters to use for C[ij]=Sum[k] A[ik]*B[jk]
-globalParameters["ScriptPath"] = os.path.dirname(os.path.realpath(__file__))            # path to Tensile/Tensile.py
-globalParameters["SourcePath"] = os.path.join(globalParameters["ScriptPath"], "Source") # path to Tensile/Source/
 globalParameters["HipClangVersion"] = "0.0.0"
 globalParameters["AMDClangVersion"] = "0.0.0"
 
@@ -1882,6 +1874,105 @@ class ProgressBar:
 
   def finish(self): pass
 
+
+class DataDirection(Enum):
+  NONE  = 0,
+  READ  = 1,
+  WRITE = 2
+
+class SpinnyThing:
+    def __init__(self):
+        self.chars = ['|', '/', '-', '\\']
+        self.index = 0;
+
+    def increment(self, value=1):
+        sys.stdout.write('\b' + self.chars[self.index])
+        sys.stdout.flush()
+        self.index = (self.index + 1) % len(self.chars)
+
+    def finish(self):
+        sys.stdout.write('\b*\n')
+        sys.stdout.flush()
+
+def iterate_progress(obj, *args, **kwargs):
+    try:
+        progress = ProgressBar(len(obj))
+    except TypeError:
+        progress = SpinnyThing()
+    for o in obj:
+        yield o
+        progress.increment()
+    progress.finish()
+
+try:
+    from tqdm import tqdm
+except ImportError:
+    tqdm = iterate_progress
+
+def state(obj):
+    if hasattr(obj, 'state'):
+        return obj.state()
+
+    if hasattr(obj.__class__, 'StateKeys'):
+        rv = {}
+        for key in obj.__class__.StateKeys:
+            attr = key
+            if isinstance(key, tuple):
+                (key, attr) = key
+            rv[key] = state(getattr(obj, attr))
+        return rv
+
+    if isinstance(obj, dict):
+        return {k: state(v) for k, v in obj.items()}
+
+    if isinstance(obj, (str, int, float)):
+        return obj
+
+    try:
+        return [state(i) for i in obj]
+    except TypeError:
+        pass
+
+    return obj
+
+def state_key_ordering(cls):
+    def tup(obj):
+        return tuple([getattr(obj, k) for k in cls.StateKeys])
+
+    def lt(a, b):
+        return tup(a) < tup(b)
+    def eq(a, b):
+        return tup(a) == tup(b)
+
+    cls.__lt__ = lt
+    cls.__eq__ = eq
+
+    return functools.total_ordering(cls)
+
+def hash_combine(*objs, **kwargs):
+    shift = 1
+    if 'shift' in kwargs:
+        shift = kwargs['shift']
+
+    if len(objs) == 1:
+        objs = objs[0]
+
+    rv = 0
+    try:
+        it = iter(objs)
+        rv = next(it)
+        for value in it:
+            rv = (rv << shift) ^ value
+    except TypeError:
+        return objs
+    except StopIteration:
+        pass
+    return rv
+
+def hash_objs(*objs, **kwargs):
+    return hash(tuple(objs))
+
+
 from copy import copy
 class Backup:
   """RAII class to restore backed up fields from object"""
@@ -1895,64 +1986,3 @@ class Backup:
     for k, v in self.fields.items():
         setattr(self.object, k, v)
 
-# Append copyrights to all files generated by tensile since they belong to Tensile intellectual property
-CMakeHeader = """################################################################################
-#
-# Copyright (C) 2022 Advanced Micro Devices, Inc. All rights reserved.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell cop-
-# ies of the Software, and to permit persons to whom the Software is furnished
-# to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IM-
-# PLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNE-
-# CTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-################################################################################
-
-###################################################
-# This file was generated by Tensile:             #
-# https://github.com/ROCmSoftwarePlatform/Tensile #
-###################################################
-
-
-"""
-
-CHeader = """/*******************************************************************************
-* Copyright (C) 2022 Advanced Micro Devices, Inc. All rights reserved.
-*
-* Permission is hereby granted, free of charge, to any person obtaining a copy
-* of this software and associated documentation files (the "Software"), to deal
-* in the Software without restriction, including without limitation the rights
-* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell cop-
-* ies of the Software, and to permit persons to whom the Software is furnished
-* to do so, subject to the following conditions:
-*
-* The above copyright notice and this permission notice shall be included in all
-* copies or substantial portions of the Software.
-*
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IM-
-* PLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-* FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-* COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-* IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNE-
-* CTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-*******************************************************************************/
-
-/**************************************************
-* This file was generated by Tensile:             *
-* https://github.com/ROCmSoftwarePlatform/Tensile *
-**************************************************/
-
-
-"""
-
-HR = "################################################################################"
