+++ tensilelite/Tensile/Common/Types.py
@@ -1,0 +1,3 @@
+from typing import Tuple
+
+IsaVersion = Tuple[int, int, int]
