+++ tensilelite/Tensile/KernelWriterActivationOnly.py
@@ -22,7 +22,7 @@
 
 from copy import deepcopy
 
-from .Common import globalParameters, CHeader
+from .Common import INDEX_CHARS
 from .KernelWriterBase import KernelWriterBase
 
 class KernelWriterActivationOnly(KernelWriterBase):
@@ -39,8 +39,8 @@ class KernelWriterActivationOnly(KernelWriterBase):
 
     # determine chars for fast access
     self.states.indexChars = []
-    for i in range(0, len(globalParameters["IndexChars"])):
-      self.states.indexChars.append(globalParameters["IndexChars"][i])
+    for i in range(0, len(INDEX_CHARS)):
+      self.states.indexChars.append(INDEX_CHARS[i])
     self.states.indexChars[self.state["ProblemType"]["Index0"]] = "0" + self.states.indexChars[self.state["ProblemType"]["Index0"]]
     self.states.indexChars[self.state["ProblemType"]["Index1"]] = "1" + self.states.indexChars[self.state["ProblemType"]["Index1"]]
     self.tileChar0 = self.states.indexChars[self.state["ProblemType"]["Index0"]]
@@ -209,7 +209,7 @@ class KernelWriterActivationOnly(KernelWriterBase):
 
 
   def getKernelName(self):
-    indexChars = globalParameters["IndexChars"]
+    indexChars = INDEX_CHARS
     # C dimensions
     name = "D"
     for i in range(0, self.state["ProblemType"]["NumIndicesC"]):
