+++ tensilelite/Tensile/KernelWriterReduction.py
@@ -24,7 +24,7 @@
 
 from copy import deepcopy
 
-from .Common import globalParameters, CHeader
+from .Common import INDEX_CHARS
 from .KernelWriterBase import KernelWriterBase
 from .TensileInstructions import DataType
 
@@ -35,12 +35,11 @@ class KernelWriterReduction(KernelWriterBase):
 
         self.state["ProblemType"] = deepcopy(state["ProblemType"])
 
-        indexChars = globalParameters["IndexChars"]
         # C dimensions
         self.indicesStr = ""
         for i in range(0, self.state["ProblemType"]["NumIndicesC"]):
-            c = indexChars[i].lower()
-            self.indicesStr += indexChars[i].lower()
+            c = INDEX_CHARS[i].lower()
+            self.indicesStr += INDEX_CHARS[i].lower()
 
         # derive parameter
         self.language = "HIP"
@@ -65,14 +64,14 @@ class KernelWriterReduction(KernelWriterBase):
 
     def getHeaderFileString(self):
         fileString = "" # CHeader
-        indexChars = globalParameters["IndexChars"]
+
         # C dimensions
         indicesStr = ""
         for i in range(0, self.state["ProblemType"]["NumIndicesC"]):
-            c = indexChars[i].lower()
+            c = INDEX_CHARS[i].lower()
             if c == 'k':
                 continue
-            indicesStr += indexChars[i].lower()
+            indicesStr += INDEX_CHARS[i].lower()
 
         computeStr  = self.state["ProblemType"]["ComputeDataType"].toDevice(self.language)
         computeChar = self.state["ProblemType"]["ComputeDataType"].toChar()
