+++ tensilelite/bump-version.sh
@@ -33,7 +33,7 @@ NEW_VERSION="4.33.0"
 OLD_MINIMUM_REQUIRED_VERSION="MinimumRequiredVersion: 4.7.2"
 NEW_MINIMUM_REQUIRED_VERSION="MinimumRequiredVersion: 4.8.0"
 
-sed -i "s/${OLD_VERSION}/${NEW_VERSION}/g" Tensile/__init__.py
+sed -i "s/${OLD_VERSION}/${NEW_VERSION}/g" Tensile/Common/Common.py
 sed -i "s/${OLD_VERSION}/${NEW_VERSION}/g" HostLibraryTests/CMakeLists.txt
 
 echo "The version number also needs to be fixed in Tensile/cmake/TensileConfigVersion.cmake ."
