+++ tensilelite/tox.ini
@@ -7,7 +7,6 @@ envlist = py35,py36,py27,lint
 # process to crash if a multiprocessing job has an exception. Fixed in 5.3.3.
 deps =
     -r{toxinidir}/requirements.txt
-    joblib
     pytest>=5.4.1
     pytest-xdist>=1.32.0
     filelock
